# docs/swagger.yaml
basePath: /api/
definitions:
...
  chat.ChatAndUserResponse:
    properties:
      chat:
        $ref: '#/definitions/models.Chat'
      user:
        $ref: '#/definitions/models.User'
    type: object
  chat.ChatIdResponse:
    properties:
      chat_id:
        type: string
    type: object
  chat.ChatListResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/models.Chat'
        type: array
    type: object
  chat.ChatResponse:
    properties:
      chat:
        $ref: '#/definitions/models.Chat'
    type: object
  chat.IdResponse:
    properties:
      id:
        type: string
    type: object
  chat.MessageResponse:
    properties:
      message:
        type: string
    type: object
  chat.NameInput:
    properties:
      name:
        type: string
    type: object
  chat.UserIdInput:
    properties:
      user_id:
        type: integer
    type: object
...
  models.Chat:
    properties:
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      types:
        type: string
    required:
    - icon
    - name
    type: object
...
  models.User:
    properties:
      icon:
        type: string
      id:
        type: integer
      password:
        type: string
      username:
        type: string
    required:
    - icon
    - password
    - username
    type: object
  responses.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    name: Tairyoku
    url: https://github.com/Tairyoku
  description: Серверна частина онлайн чату.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Server API
  version: 1.0.0
paths:
...
  /chats/{id}:
    delete:
      consumes:
      - application/json
      description: Отримує ID чату. Видаляє чат.
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: chat  deleted
          schema:
            $ref: '#/definitions/chat.MessageResponse'
        "500":
          description: messages delete error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete chat
      tags:
      - chat
    get:
      consumes:
      - application/json
      description: Отримує ID чату. Повертає дані чату.
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: result is chat data
          schema:
            $ref: '#/definitions/chat.ChatResponse'
        "500":
          description: get chat error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get chat info
      tags:
      - chat
  /chats/{id}/add:
    post:
      consumes:
      - application/json
      description: |-
        Отримує ID чату та користувача.
        Додає користувача до чату.
        Повертає ID зв'язку між чатами та користувачами.
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID
        in: body
        name: user_id
        required: true
        schema:
          $ref: '#/definitions/chat.UserIdInput'
      produces:
      - application/json
      responses:
        "200":
          description: result is ID of chats and users relations
          schema:
            $ref: '#/definitions/chat.IdResponse'
        "400":
          description: incorrect request data
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: add user to chat error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Add user to chat
      tags:
      - chat
  /chats/{id}/delete:
    put:
      consumes:
      - application/json
      description: |-
        Отримує ID чату та користувача.
        Видаляє користувача з чату.
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: integer
      - description: User ID
        in: body
        name: user_id
        required: true
        schema:
          $ref: '#/definitions/chat.UserIdInput'
      produces:
      - application/json
      responses:
        "200":
          description: user deleted from chat
          schema:
            $ref: '#/definitions/chat.MessageResponse'
        "400":
          description: incorrect request data
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: messages delete error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete user from chat
      tags:
      - chat
  /chats/{id}/icon:
    put:
      consumes:
      - application/json
      description: |-
        Отримує ID чату та файл зображення.
        Оновлює зображення користувача.
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: icon changed
          schema:
            $ref: '#/definitions/chat.MessageResponse'
        "400":
          description: incorrect chat data
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: delete icon error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Change chat icon
      tags:
      - chat
  /chats/{id}/link:
    get:
      consumes:
      - application/json
      description: |-
        Отримує ID чату.
        Повертає дані чату та, якщо чат приватний, користувача.
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: result is chat data (and user data)
          schema:
            $ref: '#/definitions/chat.ChatAndUserResponse'
        "500":
          description: get users error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get chat (and if chat is private - user) info
      tags:
      - chat
  /chats/{id}/users:
    get:
      consumes:
      - application/json
      description: Отримує ID чату. Повертає список користувачів чату.
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: result is list of chat`s users
          schema:
            items:
              items:
                $ref: '#/definitions/models.User'
              type: array
            type: array
        "500":
          description: get users error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get chat`s users
      tags:
      - chat
  /chats/{userId}/private:
    get:
      consumes:
      - application/json
      description: |-
        Отримує ID другого користувача приватного чату.
        Якщо чат вже існує, повертає його ID.
        Якщо чату не існує, створює його, та повертає ID чату.
        Якщо другий користувач є поточним користувачем,
        створює (за необхідністю) персональний чат та повертає його ID.
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: return chat ID
          schema:
            $ref: '#/definitions/chat.ChatIdResponse'
        "500":
          description: add second user to chat error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create private chat and/or get chat ID
      tags:
      - chat
  /chats/create:
    post:
      consumes:
      - application/json
      description: Отримує ім'я чату. Створює новий чат. Повертає ID чату.
      parameters:
      - description: Chat name
        in: body
        name: chat_name
        required: true
        schema:
          $ref: '#/definitions/chat.NameInput'
      produces:
      - application/json
      responses:
        "200":
          description: result is chat ID
          schema:
            $ref: '#/definitions/chat.IdResponse'
        "400":
          description: name is empty
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: add user to chat error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new public chat
      tags:
      - chat
  /chats/search/{name}:
    get:
      consumes:
      - application/json
      description: |-
        Отримує частину імені користувача.
        Повертає список користувачів, ім'я яких повністю або частково збігається.
      parameters:
      - description: Slice of chats
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: return found users
          schema:
            $ref: '#/definitions/chat.ChatListResponse'
        "500":
          description: found chats error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get found users
      tags:
      - chat
  /chats/users/{id}/private:
    get:
      consumes:
      - application/json
      description: |-
        Отримує ID користувача.
        Повертає список приватних чатів користувача.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: result is list of chats
          schema:
            $ref: '#/definitions/chat.ChatListResponse'
        "500":
          description: get users error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user`s private chats
      tags:
      - chat
  /chats/users/{id}/public:
    get:
      consumes:
      - application/json
      description: |-
        Отримує ID користувача.
        Повертає список публічних чатів, в яких є користувач.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: result is list of chats
          schema:
            $ref: '#/definitions/chat.ChatListResponse'
        "500":
          description: get chats error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get user`s public chats
      tags:
      - chat
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"